test(engine): extract the two frozen payloads to a shared module (#398)
Badminton (payload A) and Stepladder (payload B) move out of
participants-rules.test.ts into payload-fixtures.ts so #396's
participants tests and #398's instruction tests assert against
byte-identical inputs. A payload that drifted between waves would prove
nothing about either.

Also relaxes SchedulingConstraints.hard from .default([]) to .optional():
a defaulted zod field is REQUIRED in the output type, and that type is
SlotConfig.constraints, built as an object literal at dozens of call
sites. Read it as constraints?.hard ?? [].
